# docs/03_api/openapi.yaml
openapi: 3.1.0
info:
  title: Moonshot Backend
  version: 0.6.0
paths:
...
  /v1/reports/{session_id}:
    get:
      tags:
      - reports
      summary: Get Report
      operationId: get_report_v1_reports__session_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reports/{session_id}/summary:
    get:
      tags:
      - reports
      summary: Get Report Summary
      operationId: get_report_summary_v1_reports__session_id__summary_get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reports/{session_id}/human-review:
    get:
      tags:
      - reports
      summary: Get Human Review
      operationId: get_human_review_v1_reports__session_id__human_review_get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanReviewRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - reports
      summary: Put Human Review
      operationId: put_human_review_v1_reports__session_id__human_review_put
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HumanReviewUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanReviewRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reports/{session_id}/interpret:
    post:
      tags:
      - reports
      summary: Create Interpretation
      operationId: create_interpretation_v1_reports__session_id__interpret_post
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterpretationRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reports/{session_id}/interpretations/{view_id}:
    get:
      tags:
      - reports
      summary: Get Interpretation
      operationId: get_interpretation_v1_reports__session_id__interpretations__view_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      - name: view_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: View Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterpretationView'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DimensionScoreOutput:
      properties:
        key:
          type: string
          title: Key
        score:
          type: number
          title: Score
        rationale:
          type: string
          title: Rationale
        failure_modes_matched:
          items:
            type: string
          type: array
          title: Failure Modes Matched
        confidence:
          type: number
          title: Confidence
      type: object
      required:
      - key
      - score
      - rationale
      - confidence
      title: DimensionScoreOutput
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HumanReviewRecord:
      properties:
        session_id:
          type: string
          format: uuid
          title: Session Id
        tenant_id:
          type: string
          title: Tenant Id
        notes_markdown:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes Markdown
        tags:
          items:
            type: string
          type: array
          title: Tags
        override_overall_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Override Overall Score
        override_confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Override Confidence
        dimension_overrides:
          additionalProperties:
            type: number
          type: object
          title: Dimension Overrides
        reviewer_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Reviewer Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - session_id
      - tenant_id
      - created_at
      - updated_at
      title: HumanReviewRecord
    HumanReviewUpdateRequest:
      properties:
        notes_markdown:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes Markdown
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        override_overall_score:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Override Overall Score
        override_confidence:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Override Confidence
        dimension_overrides:
          anyOf:
          - additionalProperties:
              type: number
            type: object
          - type: 'null'
          title: Dimension Overrides
      type: object
      title: HumanReviewUpdateRequest
    Interpretation:
      properties:
        summary:
          type: string
          title: Summary
        suggestions:
          items:
            type: string
          type: array
          title: Suggestions
        audience:
          type: string
          title: Audience
          default: reviewer_admin_only
      type: object
      required:
      - summary
      title: Interpretation
    InterpretationRequest:
      properties:
        focus_dimensions:
          items:
            type: string
          type: array
          title: Focus Dimensions
        include_sensitivity:
          type: boolean
          title: Include Sensitivity
          default: false
        weight_overrides:
          additionalProperties:
            type: number
          type: object
          title: Weight Overrides
      type: object
      title: InterpretationRequest
    InterpretationView:
      properties:
        view_id:
          type: string
          format: uuid
          title: View Id
        session_id:
          type: string
          format: uuid
          title: Session Id
        created_at:
          type: string
          format: date-time
          title: Created At
        focus_dimensions:
          items:
            type: string
          type: array
          title: Focus Dimensions
        include_sensitivity:
          type: boolean
          title: Include Sensitivity
          default: false
        weight_overrides:
          additionalProperties:
            type: number
          type: object
          title: Weight Overrides
        breakdown:
          additionalProperties: true
          type: object
          title: Breakdown
        caveats:
          items:
            type: string
          type: array
          title: Caveats
        scoring_version_lock:
          $ref: '#/components/schemas/ScoringVersionLock'
      type: object
      required:
      - session_id
      - scoring_version_lock
      title: InterpretationView
    JobAccepted:
      properties:
        job_id:
          type: string
          format: uuid
          title: Job Id
        status:
          type: string
          title: Status
        submitted_at:
          type: string
          format: date-time
          title: Submitted At
      type: object
      required:
      - job_id
      - status
      title: JobAccepted
...
    ModelInvocationTrace:
      properties:
        provider:
          type: string
          title: Provider
        model:
          type: string
          title: Model
        prompt_hash:
          type: string
          title: Prompt Hash
        latency_ms:
          type: integer
          title: Latency Ms
      type: object
      required:
      - provider
      - model
      - prompt_hash
      - latency_ms
      title: ModelInvocationTrace
...
    Report:
      properties:
        session_id:
          type: string
          format: uuid
          title: Session Id
        score_result:
          $ref: '#/components/schemas/ScoreResult'
        interpretation:
          $ref: '#/components/schemas/Interpretation'
      type: object
      required:
      - session_id
      - score_result
      - interpretation
      title: Report
    ReportSummary:
      properties:
        session_id:
          type: string
          format: uuid
          title: Session Id
        session_status:
          type: string
          title: Session Status
        report_available:
          type: boolean
          title: Report Available
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
        needs_human_review:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Needs Human Review
        trigger_codes:
          items:
            type: string
          type: array
          title: Trigger Codes
        trigger_count:
          type: integer
          title: Trigger Count
          default: 0
        last_scored_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Scored At
        scoring_version_lock:
          anyOf:
          - $ref: '#/components/schemas/ScoringVersionLock'
          - type: 'null'
        has_human_review:
          type: boolean
          title: Has Human Review
          default: false
        final_score_source:
          anyOf:
          - type: string
            enum:
            - model
            - human_override
          - type: 'null'
          title: Final Score Source
        final_confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Final Confidence
      type: object
      required:
      - session_id
      - session_status
      - report_available
      title: ReportSummary
...
    ScoreResult:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        session_id:
          type: string
          format: uuid
          title: Session Id
        objective_metrics:
          additionalProperties: true
          type: object
          title: Objective Metrics
        dimension_scores:
          additionalProperties:
            type: number
          type: object
          title: Dimension Scores
        dimension_evidence:
          additionalProperties:
            $ref: '#/components/schemas/DimensionScoreOutput'
          type: object
          title: Dimension Evidence
        confidence:
          type: number
          title: Confidence
        needs_human_review:
          type: boolean
          title: Needs Human Review
        scorer_version:
          type: string
          title: Scorer Version
          default: 0.2.0
        rubric_version:
          type: string
          title: Rubric Version
          default: 0.1.0
        task_family_version:
          type: string
          title: Task Family Version
          default: 0.1.0
        model_hash:
          type: string
          title: Model Hash
          default: local-baseline
        llm_traces:
          items:
            $ref: '#/components/schemas/ModelInvocationTrace'
          type: array
          title: Llm Traces
        trigger_codes:
          items:
            type: string
          type: array
          title: Trigger Codes
        trigger_impacts:
          items:
            additionalProperties:
              anyOf:
              - type: number
              - type: string
            type: object
          type: array
          title: Trigger Impacts
        scored_at:
          type: string
          format: date-time
          title: Scored At
      type: object
      required:
      - session_id
      - objective_metrics
      - dimension_scores
      - confidence
      - needs_human_review
      title: ScoreResult
...
    ScoringVersionLock:
      properties:
        scorer_version:
          type: string
          title: Scorer Version
        rubric_version:
          type: string
          title: Rubric Version
        task_family_version:
          type: string
          title: Task Family Version
        model_hash:
          type: string
          title: Model Hash
      type: object
      required:
      - scorer_version
      - rubric_version
      - task_family_version
      - model_hash
      title: ScoringVersionLock
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer